Gmail Support

+++ sources/get_geckodriver.sh
@@ -1,6 +1,8 @@
 #!/bin/bash
 
-url=$(curl -s https://api.github.com/repos/mozilla/geckodriver/releases/latest | python -c "import sys, json; print(next(item['browser_download_url'] for item in json.load(sys.stdin)['assets'] if 'linux32' in item.get('browser_download_url', '')))")
-curl -s -L "$url" | tar -xz
-chmod +x geckodriver
-mv geckodriver /usr/local/bin/
+if [ ! -f /usr/local/bin/geckodriver ]; then
+    url=$(curl -s https://api.github.com/repos/mozilla/geckodriver/releases/latest | python -c "import sys, json; print(next(item['browser_download_url'] for item in json.load(sys.stdin)['assets'] if 'linux32' in item.get('browser_download_url', '')))")
+    curl -s -L "$url" | tar -xz
+    chmod +x geckodriver
+    mv geckodriver /usr/local/bin/
+fi
